Rename: toastMessage로 파일명 변경
